add type checking and modularize script.js
