+++ index.html
@@ -4,13 +4,13 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>JSGame</title>
-  <link rel="stylesheet" href="style.css">
-  <script src="script.js" defer></script>
+  <link rel="stylesheet" href="./src/styles/style.css">
+  <!-- <script src="script.js" type="module" defer></script> you can not use defer for modules -->
+  <script src="./src/js/main.js" type="module"></script>
 </head>
 <body>
   <canvas id="canvas1"></canvas>
   <img src="img/overlay.png" alt="" id="overlay">
   <img src="img/obstacles.png" alt="" id="obstacles">
-  <img src="https://interactive-examples.mdn.mozilla.net/media/cc0-images/grapefruit-slice-332-332.jpg" alt="" id="obstacles">
 </body>
 </html>
