+++ src/js/jsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "checkJs": true,
+    "target": "ES6"
+  },
+  "include": ["*"],
+}
