+++ src/styles/style.css
@@ -5,7 +5,7 @@
 }
 
 #canvas1{
-  background: url('img/background.png');
+  background: url('../../img/background.png');
   position: absolute;
   top: 50%;
   left: 50%;
